feat: add 'bin/tp'
You can now do:
  `tp ./test.js`
  `tp /path/to/test.js`
  `tp` (looks for $PWD/test.js)

You can `export NODE_BIN="..."` before running `tp` if you want to use another binary for the tests.

Also, the `startTests` method now returns a promise instead of the Runner instance. The promise resolves with the test results.

You receive a warning when the entry path does not exist, or when the runner finds zero tests.