feat: always return the `files` array for `runner.start`
Previously, if an error was thrown or the runner was stopped, the promise returned by `runner.start` would resolve with undefined, instead of an object containing the `files` array, plus other information about the test results. That object will now include a `stopped` or `error` property depending on what happens.